requirements.txt
streamlit>=1.29.0
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.0
plotly>=5.18.0
dune-client>=1.3.0
flask>=3.0.0
flask-cors>=4.0.0